Added MapManager and GameAgent scripts
MapManager holds a reference to the map data and allows agents (like players or enemies) to interact with the map. GameAgent is an abstract class that has methods that both players and enemies must implement. These methods are then used by the map manager when moving or modifying things on the map.