Added Network Support
Heed this description-
I've hardcoded in the server IP, and it should be always running, but if you run into issues with the server
you can use Network.disable(), and any commands submitted will be immediately piped back into the system, making it
effectively a single-player game. Network commands don't do anything yet, but if you check NetworkManager, there is
already a skeleton in place for handling a variety of incoming commands.